refactor(react)!: overhaul MultipleChoice as ChoiceField
This dramatically simplifies how our "choice" field elements work, allowing both the checkbox and radio fields to use a single shared component, the simplified `Choice`.

BREAKING CHANGE `MultipleChoice` has been removed in favor of the new `ChoiceField`, which will be rendered as a `<fieldset>` with either checkbox or radio choices, depending on whether the `multiple` prop is `true` or `false`.