fix(core): prevent accidental root declarations
Previously, `:root` declarations would be created any time the `@wwnds/core`
entry point was used, potentailly resulting in duplicate declarations. This
wraps all `:root` declarations in mixins to ensure that the `@wwnds/core`
entry point contains no declarations and can be used modularly without
accidentally creating duplicate declarations.

While this was the result of a bug, some applications may have been relying
on the bug for their styles, meaning this change may be breaking for those
use-cases. This will need to be documented in the release notes.